tg_new_user_watcher/requirements.txt
telebot
psycopg2-binary
